Move SyncSubmodules out of GitClient
Doesn't use libgit2sharp so GitClient didn't make sense.
