+++ src/GitHub.App/Services/GitClient.cs
@@ -3,7 +3,6 @@ using System.Collections.Generic;
 using System.ComponentModel.Composition;
 using System.IO;
 using System.Linq;
-using System.Diagnostics;
 using System.Threading.Tasks;
 using GitHub.Extensions;
 using GitHub.Models;
@@ -65,39 +64,6 @@ namespace GitHub.Services
                     repository.Network.Push(remote, "HEAD", remoteRef, pushOptions);
                 }
             });
-        }
-
-        // HACK: This is just a prototype!
-        public async Task SyncSubmodules(IRepository repository, Action<string> progress = null)
-        {
-            Guard.ArgumentNotNull(repository, nameof(repository));
-
-            var workingDir = repository.Info.WorkingDirectory;
-            var script =
-@"git submodule init
-git submodule sync --recursive
-git submodule update --recursive";
-            var scriptFile = Path.Combine(Path.GetTempPath(), "SyncSubmodules.cmd");
-            File.WriteAllText(scriptFile, script);
-
-            var startInfo = new ProcessStartInfo(scriptFile)
-            {
-                WorkingDirectory = workingDir,
-                UseShellExecute = false,
-                CreateNoWindow = true,
-                RedirectStandardOutput = true
-            };
-
-            using (var process = Process.Start(startInfo))
-            {
-                var outputReader = process.StandardOutput;
-
-                string line;
-                while ((line = await outputReader.ReadLineAsync()) != null)
-                {
-                    progress?.Invoke(line);
-                }
-            }
         }
 
         public Task Fetch(IRepository repository, string remoteName)
